developement small changes for here

+++ courses/models.py
@@ -4,7 +4,8 @@ from django.contrib.auth.models import User
 class Course(models.Model):
     title = models.CharField(max_length=200, default="Python Course")
     description = models.TextField(default="Course Description")
-    price = models.IntegerField(default=0)
+    # price = models.IntegerField(default=0)
+    price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     duration = models.CharField(max_length=50, default="3 Months")
     image = models.ImageField(upload_to='courses/', null=True, blank=True)
